sahayak-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-cloud-aiplatform==1.38.0
google-generativeai==0.3.0
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.16.3
google-cloud-translate==3.12.0
firebase-admin==6.2.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
reportlab==4.0.4
python-docx==0.8.11
openpyxl==3.1.2